// front/src/types/fullCompany.ts
export type FullCompanyType = {
  name: string;
  website: string;
  size: string;
  location: string;
  logo_url: string;
  website_url: string;
  employees_count: number;
  revenue_in_usd: number;
  industry: string;
  intent_topic: string;
  intent_dates: string[];
  about: string;
  active_wallets: number;
  token_name: string;
  tvl_usd: number;
  smart_contract_count: number;
};